requirements.txt
openai==0.27.8
streamlit==1.27.1
python-dotenv==1.0.0
chromadb==0.4.17
langchain==0.0.334
unstructured==0.10.30
torch==2.0.0
sentence-transformers==2.2.2
ipython==8.21.0